+++ src/pages/geral/geral-dados-motorista/geral-dados-motorista.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { GeralDadosMotoristaPage } from './geral-dados-motorista';
+
+
+@NgModule({
+  declarations: [
+    GeralDadosMotoristaPage,
+  ],
+  imports: [
+    IonicPageModule.forChild(GeralDadosMotoristaPage),
+  ],
+})
+export class GeralDadosMotoristaModule {}
