+++ src/pages/geral/geral-page/geral-page.html
@@ -1,0 +1,15 @@
+<ion-header>
+  <ion-navbar>
+    <ion-title>
+      Dados Gerais
+    </ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content> 
+  <ion-tabs>
+    <ion-tab [root]="geralDadosViagem" tabTitle="Viagem" tabIcon="ios-compass"></ion-tab>
+    <ion-tab [root]="geralDadosMotorista" tabTitle="Motorista" tabIcon="ios-contact"></ion-tab> 
+  </ion-tabs>
+
+</ion-content>
