+++ src/pages/geral/geral-page/geral-page.scss
