+++ src/pages/geral/geral-page/geral-page.ts
@@ -1,0 +1,21 @@
+import { Component } from '@angular/core';
+import { IonicPage, NavController, NavParams, AlertController } from 'ionic-angular';
+import { ViagensPage } from '../../viagens/viagens';
+import { GeralDadosViagemPage } from '../geral-dados-viagem/geral-dados-viagem';
+import { GeralDadosMotoristaPage } from '../geral-dados-motorista/geral-dados-motorista';
+
+ 
+
+@IonicPage()
+@Component({
+  selector: 'geral-page',
+  templateUrl: 'geral-page.html',
+})
+export class GeralPage {
+
+  constructor(public navCtrl: NavController, public navParams: NavParams, public alertCtrl: AlertController) {
+  } 
+
+  geralDadosViagem = GeralDadosViagemPage;
+  geralDadosMotorista = GeralDadosMotoristaPage;
+}
