--- src/pages/geral/geral.html
@@ -1,15 +1,0 @@
-<ion-header>
-  <ion-navbar>
-    <ion-title>
-      Dados Gerais
-    </ion-title>
-  </ion-navbar>
-</ion-header>
-
-<ion-content> 
-  <ion-tabs>
-    <ion-tab [root]="geral2" tabTitle="Viagem" tabIcon="ios-compass"></ion-tab>
-    <ion-tab [root]="geral3" tabTitle="Motorista" tabIcon="ios-contact"></ion-tab> 
-  </ion-tabs>
-
-</ion-content>
