--- src/pages/geral2/geral2.ts
@@ -1,21 +1,0 @@
-import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
-import { ViagensPage } from '../viagens/viagens';
-
-/**
- * Generated class for the Geral2Page page.
- *
- * See https://ionicframework.com/docs/components/#navigation for more info on
- * Ionic pages and navigation.
- */
-
-@IonicPage()
-@Component({
-  selector: 'page-geral2',
-  templateUrl: 'geral2.html',
-})
-export class Geral2Page {
-
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
-  } 
-}
