--- src/pages/geral3/geral3.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { Geral3Page } from './geral3';
-
-@NgModule({
-  declarations: [
-    Geral3Page,
-  ],
-  imports: [
-    IonicPageModule.forChild(Geral3Page),
-  ],
-})
-export class Geral3PageModule {}
