--- src/pages/geral3/geral3.ts
@@ -1,22 +1,0 @@
-import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
-import { ViagensPage } from '../viagens/viagens';
-
-/**
- * Generated class for the Geral3Page page.
- *
- * See https://ionicframework.com/docs/components/#navigation for more info on
- * Ionic pages and navigation.
- */
-
-@IonicPage()
-@Component({
-  selector: 'page-geral3',
-  templateUrl: 'geral3.html',
-})
-export class Geral3Page {
-
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
-  }
- 
-}
